+++ component/prometheus/exporter/dnsmasq/dnsmasq.go
@@ -2,6 +2,7 @@ package dnsmasq
 
 import (
 	"github.com/grafana/agent/component"
+	"github.com/grafana/agent/component/discovery"
 	"github.com/grafana/agent/component/prometheus/exporter"
 	"github.com/grafana/agent/pkg/integrations"
 	"github.com/grafana/agent/pkg/integrations/dnsmasq_exporter"
@@ -12,13 +13,21 @@ func init() {
 		Name:    "prometheus.exporter.dnsmasq",
 		Args:    Arguments{},
 		Exports: exporter.Exports{},
-		Build:   exporter.New(createExporter, "dnsmasq"),
+		Build:   exporter.NewWithTargetBuilder(createExporter, "dnsmasq", customizeTarget),
 	})
 }
 
 func createExporter(opts component.Options, args component.Arguments) (integrations.Integration, error) {
 	a := args.(Arguments)
 	return a.Convert().NewIntegration(opts.Logger)
+}
+
+func customizeTarget(baseTarget discovery.Target, args component.Arguments) []discovery.Target {
+	a := args.(Arguments)
+	target := baseTarget
+
+	target["instance"] = a.Address
+	return []discovery.Target{target}
 }
 
 // DefaultArguments holds the default arguments for the prometheus.exporter.dnsmasq component.