+++ CHANGELOG.md
@@ -21,6 +21,8 @@ v0.34.0-rc.1 (2023-06-02)
 
 - Fix issue where using exporters inside modules failed due to not passing the in-memory address dialer. (@mattdurham)
 
+- Fix missing `instance` key for `prometheus.exporter.dnsmasq` component. (@spartan0x117)
+
 v0.34.0-rc.0 (2023-06-01)
 -------------------------