+++ CHANGELOG.md
@@ -14,6 +14,11 @@ v0.34.0-rc.2 (2023-06-06)
 
 - `prometheus.operator.podmonitors` and `prometheus.operator.servicemonitors` can now access cluster secrets for authentication to targets. (@captncraig)
 
+### Bugfixes
+
+- Add signing region to remote.s3 component for use with custom endpoints so that Authorization Headers work correctly when
+  proxying requests. (@mattdurham)
+
 v0.34.0-rc.1 (2023-06-02)
 --------------------